requirements.txt
pytest==7.1.2
#PyMySQL==0.9.3
allure-pytest==2.8.6
requests==2.27.1
#PyYAML==5.3.1
Faker==4.0.2
selenium==4.1.3
pytest-xdist==2.5.0